old-tests/submission/Infraware/Forms/contents/Forms/textarea_dirname.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>textarea_dirname</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action="php/dirname.php"
		  id="input_form">
	<p><textarea id='textarea_id'></textarea></p>
	</form>

	<script>
		
		var textarea = document.getElementById("textarea_id");
		textarea.dirname = 'textarea_dirname';

		if (typeof(textarea.dirname) == "string") {
			test( function() {
				assert_equals(textarea.dirname, 'textarea_dirname', "dirname attribute is not correct.");
			});
		} else {
			test( function() {
				assert_unreached("dirname attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
